trying to solve an error in relation with ide

+++ src/main/java/com/example/ExpenseTracker/repository/ExpenseRepository.java
@@ -1,12 +1,9 @@
 package com.example.ExpenseTracker.repository;
 
-
 import com.example.ExpenseTracker.entity.Expense;
-import org.hibernate.event.service.spi.JpaBootstrapSensitive;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ExpenseRepository extends JpaRepository<Expense, Integer> {
-
+public interface ExpenseRepository extends JpaRepository<Expense, Long> {
 }
